+++ resources/lang/en/messages.php
@@ -2,4 +2,5 @@
 
 return [
     'matchUserPassword' => 'The given password is not your current password.',
+    'markdownHttpImageUrl' => 'The :attribute field contains at least one image with an HTTP link.',
 ];