+++ resources/lang/nl/messages.php
@@ -2,4 +2,5 @@
 
 return [
     'matchUserPassword' => 'Het gegeven wachtwoord is niet je huidig wachtwoord.',
+    'markdownHttpImageUrl' => 'Het :attribute veld bevat minstens één afbeelding met een HTTP-koppeling.',
 ];